+++ components/panics/src/memfault_fault_handling_xtensa.c
@@ -4,21 +4,59 @@
 //! See License.txt for details
 //!
 //! @brief
-//! Fault handling for Xtensa based devices
+//! Fault handling for Xtensa based architectures
 
+#include "memfault/core/platform/core.h"
+#include "memfault/panics/arch/xtensa/esp32.h"
 #include "memfault/panics/reboot_tracking.h"
 #include "memfault_reboot_tracking_private.h"
+#include "memfault/panics/coredump.h"
+#include "memfault/panics/coredump_impl.h"
+#include "memfault/core/compiler.h"
 
-#include <stdlib.h>
-
-#ifndef ESP_PLATFORM
-#  error "The current assert handler assumes the esp-idf is in use!"
-#endif
+const sMfltCoredumpRegion *memfault_coredump_get_arch_regions(size_t *num_regions) {
+  *num_regions = 0;
+  return NULL;
+}
 
 static eMfltResetReason s_crash_reason = kMfltRebootReason_Unknown;
 
-void memfault_fault_handling_assert(void *pc, void *lr, uint32_t extra) {
-  // Note: At the moment we assume the esp-idf is in use when Xtensa is being used. The SDK
-  // implements abort which will invoke the esp-idf coredump handler as well as a chip reboot.
-  abort();
+void memfault_fault_handler(const sMfltRegState *regs, eMfltResetReason reason) {
+  if (s_crash_reason == kMfltRebootReason_Unknown) {
+    sMfltRebootTrackingRegInfo info = {
+      .pc = regs->pc,
+    };
+    memfault_reboot_tracking_mark_reset_imminent(reason, &info);
+    s_crash_reason = reason;
+  }
+
+  sMemfaultCoredumpSaveInfo save_info = {
+    .regs = regs,
+    .regs_size = sizeof(*regs),
+    .trace_reason = s_crash_reason,
+  };
+
+  // Check out "Windowed Procedure-Call Protocol" in the Xtensa ISA Reference Manual: Some data is
+  // stored "below the stack pointer (since they are infrequently referenced), leaving the limited
+  // range of the ISA’s load/store offsets available for more frequently referenced locals."
+  //
+  // Processor saves callers a0..a3 in the 16 bytes below the "sp"
+  // The next 48 bytes beneath that are from a _WindowOverflow12 on exception
+  // capturing callers a4 - a15
+  //
+  // For the windowed ABI, a1 always holds the current "sp":
+  //   https://github.com/espressif/esp-idf/blob/v4.0/components/freertos/readme_xtensa.txt#L421-L428
+  const uint32_t windowed_abi_spill_size = 64;
+  const uint32_t sp_prior_to_exception = regs->a[1] - windowed_abi_spill_size;
+
+  sCoredumpCrashInfo info = {
+    .stack_address = (void *)sp_prior_to_exception,
+    .trace_reason = save_info.trace_reason,
+  };
+  save_info.regions = memfault_platform_coredump_get_regions(&info, &save_info.num_regions);
+
+  const bool coredump_saved = memfault_coredump_save(&save_info);
+  if (coredump_saved) {
+    memfault_reboot_tracking_mark_coredump_saved();
+  }
 }