+++ ports/esp_idf/memfault/include/memfault/esp_port/cli.h
@@ -1,0 +1,11 @@
+#pragma once
+
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//! ESP32 Memfault Demo Cli Port
+
+//! Call on boot to enabled the mflt demo cli
+//! See https://mflt.io/demo-cli for more info
+void memfault_register_cli(void);