+++ ports/esp_idf/memfault/include/v4.x/memfault/esp_port/uart.h
@@ -1,0 +1,13 @@
+#pragma once
+
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//!
+//! @brief
+//! Thin wrapper around rom/uart which was refactored between the v3.x and v4.x esp-idf releases
+
+#include "esp32/rom/uart.h"
+
+#define MEMFAULT_ESP32_CONSOLE_UART_NUM CONFIG_ESP_CONSOLE_UART_NUM